Ensure regex matches the exact tag.
Currently, the regex will match the beginning of the tag (e.g. `<mj-image*`), instead of looking for an exact match. This works fine if you're using the default MJML components. However, if you have a custom component, like `mj-image-variable`, it will look for `<mj-image*`, which means it'll match `<mj-image-variable`, causing issues.